Updated components with JDOCs and minor changes
