+++ frontend/utils/MiscellaneousUtils.js
@@ -213,6 +213,120 @@ function truncateString(str, num, suffix = '...') {
   return str.slice(0, num - suffix.length) + suffix;
 }
 
+/**
+ * Copies content to the clipboard.
+ *
+ * @param {Object} params - The parameters containing the content to copy.
+ * @param {string} params.createdDate - The date the content was created.
+ * @param {string} params.title - The title of the content.
+ * @param {string} params.description - The description of the content.
+ * @param {Array} params.multipleChoiceList - List of multiple choice questions and answers.
+ * @param {Array} params.flashCards - List of flash cards with concepts and definitions.
+ * @return {Promise<void>} A promise that resolves when the content is copied to the clipboard.
+ */
+const copyContentToClipboard = ({
+  createdDate,
+  title,
+  description,
+  multipleChoiceList,
+  flashCards,
+}) => {
+  return new Promise((resolve, reject) => {
+    let content = `Created Date: ${createdDate}\n`;
+    content += `Title: ${title}\n`;
+    content += `Description: ${description}\n\n`;
+
+    // Append Multiple Choice Questions and Answers
+    if (multipleChoiceList && multipleChoiceList.length > 0) {
+      content += 'Questions and Options:\n';
+      multipleChoiceList.forEach((item, index) => {
+        content += `${index + 1}. ${item.question}\n`;
+        item.choices.forEach((choice, qIndex) => {
+          content += `   ${String.fromCharCode(65 + qIndex)}. ${
+            choice.value
+          }\n`;
+        });
+        content += '\n';
+        content += `Answer ${index + 1}: ${String.fromCharCode(
+          65 + item.answer.charCodeAt(0) - 'A'.charCodeAt(0)
+        )}\n`;
+        content += `Explanation: ${item.explanation}\n\n`;
+      });
+    }
+
+    // Append Flash Cards
+    if (flashCards && flashCards.length > 0) {
+      content += 'Flash Cards:\n';
+      flashCards.forEach((card, index) => {
+        content += `${index + 1}. ${card.concept}\n`;
+        content += `   - ${card.definition}\n\n`;
+      });
+    }
+
+    navigator.clipboard.writeText(content).then(resolve, reject);
+  });
+};
+
+/**
+ * Exports content as a CSV file.
+ *
+ * @param {Object} params - The parameters containing the content to export.
+ * @param {string} params.createdDate - The date the content was created.
+ * @param {string} params.title - The title of the content.
+ * @param {string} params.description - The description of the content.
+ * @param {Array} params.multipleChoiceList - List of multiple choice questions and answers.
+ * @param {Array} params.flashCards - List of flash cards with concepts and definitions.
+ */
+const exportContentAsCSV = ({
+  createdDate,
+  title,
+  description,
+  multipleChoiceList,
+  flashCards,
+}) => {
+  let csvContent = '"Created Date","Title","Description"\n';
+  csvContent += `"${createdDate}","${title}","${description}"\n\n`;
+
+  // Export Multiple Choice List
+  if (multipleChoiceList && multipleChoiceList.length > 0) {
+    multipleChoiceList.forEach((item, index) => {
+      csvContent += `"Question ${index + 1}","${item.question}","\n`;
+
+      item.choices.forEach((choice, choiceIndex) => {
+        csvContent += `"","Option ${String.fromCharCode(65 + choiceIndex)}","${
+          choice.value
+        }"\n`;
+      });
+
+      csvContent += '"\n';
+
+      csvContent += `"Answer ${index + 1}","${String.fromCharCode(
+        65 + item.answer.charCodeAt(0) - 'A'.charCodeAt(0)
+      )}","${item.explanation}"\n\n`;
+    });
+  }
+
+  // Export Flash Cards
+  if (flashCards && flashCards.length > 0) {
+    csvContent += '\n\n"Concept","Definition"\n';
+    flashCards.forEach((card) => {
+      csvContent += `"${card.concept}","${card.definition}"\n`;
+    });
+  }
+
+  const fileName = `${title
+    .replace(/ /g, '_')
+    .toLowerCase()}_${createdDate}.csv`;
+
+  const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = fileName;
+  a.click();
+  URL.revokeObjectURL(url);
+};
+
 export {
   toBase64,
   shimmerEffect,
@@ -223,4 +337,6 @@ export {
   getRandomBackgroundColor,
   disableScroll,
   truncateString,
+  copyContentToClipboard,
+  exportContentAsCSV,
 };
